WIP Fix: App-Core arch

+++ core/application/layer.hpp
@@ -6,7 +6,7 @@ public:
     virtual ~ILayer() = default;
     virtual void OnAttach() = 0;
     virtual void OnDetach() = 0;
-    virtual void OnUpdate(float deltaTime) = 0;
+    virtual void OnUpdate(double deltaTime) = 0;
     virtual void OnRender() = 0;
 };
 }
